+++ ribs/emitters/_evolution_strategy_emitter.py
@@ -109,7 +109,7 @@ class EvolutionStrategyEmitter(EmitterBase):
         self._ranker.reset(self, archive, self._rng)
 
         self._batch_size = self.opt.batch_size
-        self._restarts = 0  # Currently not exposed publicly.
+        self._restarts = 0
 
     @property
     def x0(self):
@@ -120,6 +120,11 @@ class EvolutionStrategyEmitter(EmitterBase):
     def batch_size(self):
         """int: Number of solutions to return in :meth:`ask`."""
         return self._batch_size
+
+    @property
+    def restarts(self):
+        """int: The number of restarts for this emitter."""
+        return self._restarts
 
     def ask(self):
         """Samples new solutions from a multivariate Gaussian.
@@ -151,13 +156,12 @@ class EvolutionStrategyEmitter(EmitterBase):
              metadata_batch=None):
         """Gives the emitter results from evaluating solutions.
 
-        As we insert solutions into the archive, we record the solutions'
-        impact on the fitness of the archive. For example, if the added
-        solution makes an improvement on an existing elite, then we
-        will record ``(AddStatus.IMPROVED_EXISTING, improvement_value)``
-
         The solutions are ranked based on the `rank()` function defined by
-        `self._ranker`.
+        `self._ranker`. Then, the ranked solutions are passed to CMA-ES for
+        adaptation.
+
+        This function also checks for restart condition and restarts CMA-ES
+        when needed.
 
         Args:
             solution_batch (numpy.ndarray): (batch_size, :attr:`solution_dim`)